Add database setup scripts and docker files
Add a Dockerfile and docker-compose.yml file for spinning up a development PostgreSQL database, add a databaseSetup SQL file that the Dockerfile copies in and runs on initialization of the development database. Add a database.js file that has two functions for now, initialization of the database by running the createTables.sql file, and seed which runs the insertSampleMessages.sql file to seed the database table with sample messages. For the database.js functions I added npm scripts to run them in package.json, like npm run initialize, and also added libraries used in database.js for promises and talking to the database, bluebird and pg-promise.