Add new UI to get messages and display them
Add MaterialUI library to use Card and GridV2 components for new UI.
Create MessageCard component which takes a message and returns a component with the message id and text as a card.
Create MessageList compnent which uses ServerRequests to get messages from the server and maps them to the MessageCard component to display in a list.
Add new MessageList component to App.tsx and update App.css to set the width of div with classname App to half the view width.